+++ src/components/pages/Admin/components/RegistrationRequests/components/RequestList/index.tsx
@@ -37,6 +37,11 @@ import type { GridCellParams } from '@mui/x-data-grid'
 import './RequestListStyle.scss'
 import { refetch } from 'features/admin/registration/slice'
 
+interface fetchHookArgsType {
+  statusFilter: string
+  expr: string
+}
+
 export const RequestList = ({
   fetchHook,
   onTableCellClick,
@@ -68,7 +73,7 @@ export const RequestList = ({
   const [filterStatus, setFilterStatus] = useState<string>(
     AppFilterType.INREVIEW
   )
-  const [fetchHookArgs, setFetchHookArgs] = useState({})
+  const [fetchHookArgs, setFetchHookArgs] = useState<fetchHookArgsType>()
   const setView = (e: React.MouseEvent<HTMLInputElement>) => {
     const viewValue = e.currentTarget.value
     setFilterStatus(viewValue)
@@ -128,7 +133,7 @@ export const RequestList = ({
 
   return (
     <section id="registration-section-id">
-      <PageLoadingTable<ApplicationRequest>
+      <PageLoadingTable<ApplicationRequest, fetchHookArgsType>
         searchExpr={searchExpr}
         rowHeight={group !== AppFilterType.CLOSED ? 160 : 80}
         alignCell="start"