+++ src/components/pages/EdcConnector/index.tsx
@@ -26,8 +26,8 @@ import UserService from 'services/UserService'
 import {
   PageHeader,
   Typography,
-  PageLoadingTable,
   PageSnackbar,
+  PageLoadingTable,
 } from '@catena-x/portal-shared-components'
 import PictureWithText from 'components/shared/frame/PictureWithText'
 import AddConnectorOverlay from './AddConnectorOverlay'
@@ -280,7 +280,7 @@ const EdcConnector = () => {
         />
       </section>
       <div className="connector-table-container">
-        <PageLoadingTable<ConnectorResponseBody>
+        <PageLoadingTable<ConnectorResponseBody, {}>
           toolbarVariant="premium"
           title={t('content.edcconnector.tabletitle')}
           loadLabel={t('global.actions.more')}
@@ -292,7 +292,7 @@ const EdcConnector = () => {
         />
       </div>
       <div className="connector-table-container">
-        <PageLoadingTable<ConnectorResponseBody>
+        <PageLoadingTable<ConnectorResponseBody, {}>
           toolbarVariant="premium"
           title={t('content.edcconnector.managedtabletitle')}
           loadLabel={t('global.actions.more')}