+++ src/features/newPartnerNetwork/partnerNetworkApiSlice.ts
@@ -24,9 +24,9 @@ import type {
   PaginResult,
   PaginFetchArgs,
 } from '@catena-x/portal-shared-components'
-import type { BusinessPartnerSearchResponse } from './types'
+import type { BusinessPartner } from './types'
 import Patterns from 'types/Patterns'
-import type { BusinessPartnerResponse } from 'features/partnerNetwork/types'
+import type { BusinessPartnerAddressResponse } from 'features/partnerNetwork/types'
 
 const checkIfBPNLNumber = (text: string): boolean =>
   Patterns.BPN.test(text.trim())
@@ -36,7 +36,7 @@ export const apiSlice = createApi({
   baseQuery: fetchBaseQuery(apiBpdmQuery()),
   endpoints: (builder) => ({
     fetchBusinessPartnerAddress: builder.mutation<
-      PaginResult<BusinessPartnerResponse>,
+      PaginResult<BusinessPartnerAddressResponse>,
       any
     >({
       query: (arry) => ({
@@ -46,7 +46,7 @@ export const apiSlice = createApi({
       }),
     }),
     fetchBusinessPartners: builder.query<
-      PaginResult<BusinessPartnerSearchResponse>,
+      PaginResult<BusinessPartner>,
       PaginFetchArgs
     >({
       query: (fetchArgs) => {
@@ -58,6 +58,29 @@ export const apiSlice = createApi({
           return `/catena/legal-entities/${fetchArgs.args!.expr}`
         } else {
           return `/catena/legal-entities?page=${fetchArgs.page}&size=10`
+        }
+      },
+      transformResponse: (response: any) => {
+        if (response.content) {
+          return {
+            ...response,
+            meta: {
+              contentSize: response.contentSize,
+              page: response.page,
+              totalElements: response.totalElements,
+              totalPages: response.totalPages,
+            },
+          }
+        } else {
+          return {
+            content: [response],
+            meta: {
+              contentSize: response.contentSize,
+              page: response.page,
+              totalElements: response.totalElements,
+              totalPages: response.totalPages,
+            },
+          }
         }
       },
     }),